+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 
 ### Changes
 - If resources in Neptune.ai already exist (when creating) or do not exist (when deleting) - only warn ([#3](https://github.com/AleksanderWWW/terraform-provider-neptune/pull/3))
+- Read config with Viper ([#5](https://github.com/AleksanderWWW/terraform-provider-neptune/pull/5))